增加单选日历

+++ liuhui/src/main/java/cn/world/liuhui/utils/DensityUtil.java
@@ -11,7 +11,7 @@ public class DensityUtil {
     /**
      * dp转换px
      */
-    public static int dp2px(Context context,int dip) {
+    public static int dp2px(Context context,float dip) {
         final float scale = context.getResources().getDisplayMetrics().density;
         return (int) (dip * scale + 0.5f);
     }
